SL-81 Moved deserializing-related interfaces to types.ts
