--- packages/http/src/validator/deserializer/IHttpHeaderParamStyleDeserializer.ts
@@ -1,6 +1,0 @@
-export type DeserializeHttpHeader = (value: string, type: string, explode: boolean) => any;
-
-export interface IHttpHeaderParamStyleDeserializer {
-  supports: (style: string) => boolean;
-  deserialize: DeserializeHttpHeader;
-}
