--- packages/http/src/validator/deserializer/IHttpParamDeserializerRegistry.ts
@@ -1,3 +1,0 @@
-export interface IHttpParamDeserializerRegistry<T extends Function> {
-  get(style: string): T | undefined;
-}
