--- packages/http/src/validator/deserializer/IHttpParamStyleDeserializer.ts
@@ -1,4 +1,0 @@
-export interface IHttpParamStyleDeserializer<T> {
-  supports: (style: string) => boolean;
-  deserialize: T;
-}
