--- packages/http/src/validator/deserializer/IHttpQueryParamStyleDeserializer.ts
@@ -1,17 +1,0 @@
-import { ISchema } from '@stoplight/types/schema';
-
-import { IHttpParamStyleDeserializer } from './IHttpParamStyleDeserializer';
-
-export type DeserializeHttpQuery = (
-  key: string,
-  query: {
-    [name: string]: string | string[];
-  },
-  schema: ISchema,
-  explode: boolean
-) => any;
-
-export interface IHttpQueryParamStyleDeserializer
-  extends IHttpParamStyleDeserializer<DeserializeHttpQuery> {
-  deserialize: DeserializeHttpQuery;
-}
